+++ src/pytest_mock_resources/__init__.py
@@ -2,15 +2,15 @@
 from pytest_mock_resources.container import (
     _mongo_container,
     _mysql_container,
-    _postgres_container,
     _redis_container,
-    _redshift_container,
     MongoConfig,
     MysqlConfig,
     PostgresConfig,
     RedisConfig,
 )
 from pytest_mock_resources.fixture.database import (
+    _postgres_container,
+    _redshift_container,
     create_mongo_fixture,
     create_mysql_fixture,
     create_postgres_fixture,
@@ -21,6 +21,7 @@ from pytest_mock_resources.fixture.database import (
     pmr_mysql_config,
     pmr_postgres_config,
     pmr_redis_config,
+    pmr_redshift_config,
     Rows,
     Statements,
 )