+++ src/pytest_mock_resources/container/__init__.py
@@ -1,6 +1,5 @@
 # flake8: noqa
 from pytest_mock_resources.container.mongo import _mongo_container, MongoConfig
 from pytest_mock_resources.container.mysql import _mysql_container, MysqlConfig
-from pytest_mock_resources.container.postgres import _postgres_container, PostgresConfig
+from pytest_mock_resources.container.postgres import PostgresConfig
 from pytest_mock_resources.container.redis import _redis_container, RedisConfig
-from pytest_mock_resources.container.redshift import _redshift_container